remove swipeToDelete
after experimentation it's way too easy to accidentally delete a set by swiping. Also, the swiping is not really necessary for notes with the redesign since the delete buttons do not take any extra space
